feat: add left join in view cart

+++ internal/entity/cart_item.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type CartItem struct {
-	ID        int       `json:"id" db:"id"`
-	CartID    int       `json:"cart_id" db:"cart_id"`
+	ID        int64     `json:"id" db:"id"`
+	CartID    int64     `json:"cart_id" db:"cart_id"`
 	Product   string    `json:"product" db:"product"`
 	Price     float64   `json:"price" db:"price"`
 	CreatedAt time.Time `json:"created_at" db:"created_at"`
